+++ libs/components/modals/src/lib/modules/modal/modal-host.component.ts
@@ -60,6 +60,8 @@ export class SkyModalHostComponent implements OnDestroy {
 
   #modalHostContext: SkyModalHostContext;
 
+  #modalInstances: SkyModalInstance[] = [];
+
   constructor(
     resolver: ComponentFactoryResolver,
     adapter: SkyModalAdapterService,
@@ -77,7 +79,9 @@ export class SkyModalHostComponent implements OnDestroy {
   }
 
   public ngOnDestroy(): void {
-    this.#modalHostContext.teardownCallback();
+    // Close all modal instances before disposing of the host container.
+    this.#closeAllModalInstances();
+    this.#modalHostContext.args.teardownCallback();
   }
 
   public open(
@@ -137,6 +141,8 @@ export class SkyModalHostComponent implements OnDestroy {
 
     modalInstance.componentInstance = modalComponentRef.instance;
 
+    this.#registerModalInstance(modalInstance);
+
     function closeModal() {
       hostService.destroy();
       adapter.setPageScroll(SkyModalHostService.openModalCount > 0);
@@ -168,10 +174,25 @@ export class SkyModalHostComponent implements OnDestroy {
 
     modalInstance.closed.subscribe(() => {
       isOpen = false;
+      this.#unregisterModalInstance(modalInstance);
       closeModal();
     });
 
     // Necessary if the host was created via a consumer's lifecycle hook such as ngOnInit
     this.#changeDetector.detectChanges();
+  }
+
+  #registerModalInstance(instance: SkyModalInstance): void {
+    this.#modalInstances.push(instance);
+  }
+
+  #unregisterModalInstance(instance: SkyModalInstance): void {
+    this.#modalInstances.slice(this.#modalInstances.indexOf(instance), 1);
+  }
+
+  #closeAllModalInstances(): void {
+    for (const instance of this.#modalInstances) {
+      instance.close();
+    }
   }
 }